Extract setup and teardown for tmux tests
This was long overdue since the setup logic is much more complex than
the actual tests.

tmux-prompt.fish had logic to protect against XDG_CONFIG_HOME with
leading double double-dot.  I believe this is no longer necessary
with the new test driver.

We still do our own setup and cleanup because we want to be able to
run this independently of the test driver,
This can be useful for debugging tests.  For example we can insert a
"$tmux attach" somewhere, and then run

	build/fish -C 'source tests/test_functions/isolated-tmux.fish' tests/checks/tmux-bind.fish

This allows to inspect the state of the test and debug interactively.
Attaching to the terminal doesn't work when running inside littlecheck
because littlecheck consumes our output and doesn't give us a terminal.
(Maybe there's an easy way to fix that?)