Fix error checks for repeated quoted command substitution
Commit e40eba358 (Treat text following quoted command substitution
as quoted) made parse_util_locate_cmdsubst_range() aware of quoted
command substitutions, by simply skipping them via quote_end().
For example «"$()this-is-skipped-until-the-quote"», we are at the final
double quote after one call to parse_util_locate_cmdsubst_range().

Now in "$()this-is-skipped$()", the $() behave as if they were "$()".
(We do not actually insert the double quotes -- only during expansion).
So when we call parse_util_locate_cmdsubst_range() in a loop, after
the first time we have $()" remaining. This is bad because the call
in the next loop iteration does not see that $() is quoted.

Fix this by inserting the double quote, just like the expansion does.

Fixes #8500