reader: when updating commandline, also update rendered highlighting
Whenever the command line changes, we redraw it with the previous syntax
highlighting. At the same time we recomputing highlighting in a background
thread.

On some systems, we are slow to recompute highlighting, so the stale syntax
highlighting is displayed for a perceptible amount.

The stale highlighting was computed for an old commandline.  When the user
had inserted or deleted some characters in the middle, then it is wrong for
all characters to the right.  This is because the characters to the right have
shifted but the highlighting hasn't.  Fix this by also shifting highlighting.

This means that text that was alrady highlighted will use the same
highlighting until a new one is computed. Newly inserted text uses the color
left of the cursor.

This is implemented by giving editable_line_t ownership of the highlighting.
It is able to perfectly sync text and highlighting; they will invariably
have the same length.

Fixes #9180