Start adding proper waiting to tmux tests
The following assumes that

	tmux send-keys a
	tmux send-keys b

are always processed in this order by the tmux server.
(If that's not true, we can spawn a persistent tmux client in control
mode to send both keys).

However the actual processing of keys (by tmux/fish etc.) is async,
which motivates a lot of sleeps.  Replace the sleeps with a synchronous
command, that only returns once fish has finished executing the
command.

The synchronization with fish is done via the input queue, by sending
a special fish-semaphore-post input sequence.  This leads to the
problem that sometimes the input shows up in the terminal (if fish
has not yet turned off ECHO.  Try to hack around this with "string replace".
Not sure if this is robust.

Another problem is that

	tests/test_driver.py target/debug

sometimes fails.  This is probably because when we run "tmux
capture-pane -p", either fish or tmux hasn't actually rendered the
latest state.  We could probably fix this by adding a test-only command
to fish that does the same thing as fish-semaphore-post but only
after forcing a render.  (and we may need something similar for tmux.

While at it, unify all remaining sleeps/timeout.